# ColdBlockInfo: overhaul analysis pass