--- a/include/swift/SILOptimizer/Analysis/ColdBlockInfo.h
+++ b/include/swift/SILOptimizer/Analysis/ColdBlockInfo.h
@@ -15,52 +15,72 @@
 
 #include "llvm/ADT/DenseMap.h"
 #include "swift/SIL/SILValue.h"
+#include "swift/Basic/FixedBitSet.h"
 
 namespace swift {
 class DominanceAnalysis;
+class PostDominanceAnalysis;
 class SILBasicBlock;
+class CondBranchInst;
 
 /// Cache a set of basic blocks that have been determined to be cold or hot.
 ///
 /// This does not inherit from SILAnalysis because it is not worth preserving
 /// across passes.
 class ColdBlockInfo {
-  DominanceAnalysis *DA;
+public:
+  // Represents the temperatures of edges flowing into a block.
+  //
+  //         T = "top" -- both warm and cold edges
+  //        /  \
+  //     Warm  Cold
+  //        \  /
+  //         B = "bottom" -- neither warm or cold edges
+  struct State {
+    // Each state kind, as an integer, is its position in any bit vectors.
+    enum Temperature {
+      Warm = 0,
+      Cold = 1
+    };
 
-  /// Each block in this map has been determined to be either cold or hot.
-  llvm::DenseMap<const SILBasicBlock*, bool> ColdBlockMap;
+    // Number of states, excluding Top or Bottom, in this flow problem.
+    static constexpr unsigned NumStates = 2;
+  };
+
+  using Energy = FixedBitSet<State::NumStates, State::Temperature>;
+
+private:
+  DominanceAnalysis *DA;
+  PostDominanceAnalysis *PDA;
+
+  /// Each block in this map has been determined to be cold and/or warm.
+  llvm::DenseMap<const SILBasicBlock*, Energy> EnergyMap;
+  bool changedMap = false;
 
   // This is a cache and shouldn't be copied around.
   ColdBlockInfo(const ColdBlockInfo &) = delete;
   ColdBlockInfo &operator=(const ColdBlockInfo &) = delete;
+  LLVM_DUMP_METHOD void dump() const;
 
-  /// Tri-value return code for checking branch hints.
-  enum BranchHint : unsigned {
-    None,
-    LikelyTrue,
-    LikelyFalse
-  };
+  void resetToCold(const SILBasicBlock *BB);
+  void set(const SILBasicBlock *BB, State::Temperature temp);
 
-  enum {
-    RecursionDepthLimit = 3
-  };
+  using ExpectedValue = std::optional<bool>;
 
-  BranchHint getBranchHint(SILValue Cond, int recursionDepth);
+  void setExpectedCondition(CondBranchInst *CondBranch, ExpectedValue value);
 
-  bool isSlowPath(const SILBasicBlock *FromBB, const SILBasicBlock *ToBB,
-                  int recursionDepth);
+  ExpectedValue searchForExpectedValue(SILValue Cond,
+                                       unsigned recursionDepth = 0);
+  void searchForExpectedValue(CondBranchInst *CondBranch);
 
-  bool isCold(const SILBasicBlock *BB,
-              int recursionDepth);
+  bool inferFromEdgeProfile(SILBasicBlock *BB);
 
 public:
-  ColdBlockInfo(DominanceAnalysis *DA): DA(DA) {}
+  ColdBlockInfo(DominanceAnalysis *DA, PostDominanceAnalysis *PDA);
 
-  bool isSlowPath(const SILBasicBlock *FromBB, const SILBasicBlock *ToBB) {
-    return isSlowPath(FromBB, ToBB, 0);
-  }
+  void analyze(SILFunction *F);
 
-  bool isCold(const SILBasicBlock *BB) { return isCold(BB, 0); }
+  bool isCold(const SILBasicBlock *BB) const;
 };
 } // end namespace swift