… lib/SILOptimizer/IPO/CapturePropagation.cpp
… lib/SILOptimizer/LoopTransforms/COWArrayOpt.cpp
… lib/SILOptimizer/Transforms/PerformanceInliner.cpp
… test/SILOptimizer/cold_block_info.swift
… test/SILOptimizer/inliner_coldblocks.sil