@kavon 2024-08-26
@swift-ci Please test compiler performance