@kavon 2024-08-27
### Performance (x86_64): -O

**Regression**                           | **OLD**  | **NEW**  | **DELTA** | **RATIO**
:---                                     | ---:     | ---:     | ---:   | ---:     
FlattenListLoop                          | 1086.0   | 1568.0   | +44.4% | **0.69x (?)**
Array.removeAll.keepingCapacity.Object   | 5.899    | 7.444    | +26.2% | **0.79x (?)**
ArrayAppendGenericStructs                | 1450.0   | 1810.0   | +24.8% | **0.80x (?)**
AngryPhonebook.Strasse.Small             | 143.5    | 167.091  | +16.4% | **0.86x (?)**
FlattenListFlatMap                       | 3808.0   | 4342.0   | +14.0% | **0.88x (?)**
StringHashing_zalgo                      | 2182.813 | 2426.563 | +11.2% | **0.90x (?)**
StringComparison_zalgo                   | 37575.0  | 41725.0  | +11.0% | **0.90x (?)**
UTF8Decode_InitFromData_ascii_as_ascii   | 167.5    | 184.083  | +9.9%  | **0.91x (?)**
String.initRepeating.longMixStr.Count100 | 706.5    | 767.0    | +8.6%  | **0.92x (?)**
AngryPhonebook.Cyrillic.Small            | 185.917  | 200.545  | +7.9%  | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                          | **OLD**  | **NEW**  | **DELTA** | **RATIO**
Data.hash.Medium                         | 28.533   | 24.276   | -14.9% | **1.18x (?)**
ArrayAppendToGeneric                     | 338.571  | 293.729  | -13.2% | **1.15x (?)**

### Code size: -O


### Performance (x86_64): -Osize

**Regression**               | **OLD**  | **NEW**  | **DELTA** | **RATIO**
:---                         | ---:     | ---:     | ---:   | ---:     
AngryPhonebook.Strasse.Small | 143.615  | 166.364  | +15.8% | **0.86x (?)**
SubstringEquatable           | 308.333  | 340.333  | +10.4% | **0.91x (?)**
StringComparison_zalgo       | 37550.0  | 41150.0  | +9.6%  | **0.91x (?)**
StringHashing_zalgo          | 2196.875 | 2366.071 | +7.7%  | **0.93x (?)**
&nbsp; | | | | 
**Improvement**              | **OLD**  | **NEW**  | **DELTA** | **RATIO**
FlattenListLoop              | 1561.0   | 1030.0   | -34.0% | **1.52x (?)**
Data.hash.Medium             | 28.523   | 24.273   | -14.9% | **1.18x (?)**
FlattenListFlatMap           | 4305.0   | 3842.0   | -10.8% | **1.12x (?)**

### Code size: -Osize

**Improvement**       | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                  | ---: | ---: | ---:  | ---:   
CharacterRecognizer.o | 4539 | 4213 | -7.2% | **1.08x**

### Performance (x86_64): -Onone

**Regression**                   | **OLD**  | **NEW**  | **DELTA** | **RATIO**
:---                             | ---:     | ---:     | ---:   | ---:     
CxxStringConversion.cxx.to.swift | 126.5    | 150.0    | +18.6% | **0.84x (?)**
&nbsp; | | | | 
**Improvement**                  | **OLD**  | **NEW**  | **DELTA** | **RATIO**
Data.hash.Medium                 | 32.87    | 28.792   | -12.4% | **1.14x (?)**
DataToStringEmpty                | 1119.048 | 1011.364 | -9.6%  | **1.11x (?)**
SIMDReduce.Int8                  | 7339.0   | 6735.0   | -8.2%  | **1.09x (?)**

### Code size: -swiftlibs



<details>
  <summary><strong>How to read the data</strong></summary>
The tables contain differences in performance which are larger than 8% and
differences in code size which are larger than 1%.

If you see any unexpected regressions, you should consider fixing the
regressions before you merge the PR.

**Noise**: Sometimes the performance results (not code size!) contain false
alarms. Unexpected regressions which are marked with '(?)' are probably noise.
If you see regressions which you cannot explain you can try to run the
benchmarks again. If regressions still show up, please consult with the
performance team (@eeckstein).
</details>

<details>
  <summary><strong>Hardware Overview</strong></summary>

      Model Name: Mac mini
      Model Identifier: Macmini8,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 3.2 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Memory: 32 GB

</details>