@kavon 2024-08-29
@swift-ci smoke test Linux