@kavon 2024-08-29
@swift-ci Please smoke test compiler performance