@kavon 2024-08-30
@swift-ci Please smoke test compiler performance