@kavon 2024-08-30
### Performance (x86_64): -O

**Regression**     | **OLD** | **NEW** | **DELTA** | **RATIO**
:---               | ---:   | ---:   | ---:   | ---:     
FlattenListFlatMap | 4097.0 | 4841.0 | +18.2% | **0.85x (?)**

### Code size: -O


### Performance (x86_64): -Osize

**Improvement**       | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                  | ---:   | ---:   | ---:   | ---:   
ReversedBidirectional | 7625.0 | 6135.0 | -19.5% | **1.24x**

### Code size: -Osize

**Improvement**       | **OLD** | **NEW** | **DELTA** | **RATIO**
:---                  | ---: | ---: | ---:  | ---:   
ReversedCollections.o | 7193 | 6818 | -5.2% | **1.06x**

### Performance (x86_64): -Onone

**Regression**                        | **OLD**  | **NEW** | **DELTA** | **RATIO**
:---                                  | ---:     | ---:   | ---:   | ---:     
NSStringConversion.InlineBuffer.ASCII | 5014.0   | 5393.0 | +7.6%  | **0.93x (?)**
&nbsp; | | | | 
**Improvement**                       | **OLD**  | **NEW** | **DELTA** | **RATIO**
ArrayAppendGenericStructs             | 1636.667 | 860.0  | -47.5% | **1.90x (?)**
CharacterLiteralsLarge                | 384.8    | 341.5  | -11.3% | **1.13x (?)**

### Code size: -swiftlibs



<details>
  <summary><strong>How to read the data</strong></summary>
The tables contain differences in performance which are larger than 8% and
differences in code size which are larger than 1%.

If you see any unexpected regressions, you should consider fixing the
regressions before you merge the PR.

**Noise**: Sometimes the performance results (not code size!) contain false
alarms. Unexpected regressions which are marked with '(?)' are probably noise.
If you see regressions which you cannot explain you can try to run the
benchmarks again. If regressions still show up, please consult with the
performance team (@eeckstein).
</details>

<details>
  <summary><strong>Hardware Overview</strong></summary>

      Model Name: Mac mini
      Model Identifier: Macmini8,1
      Processor Name: 6-Core Intel Core i7
      Processor Speed: 3.2 GHz
      Number of Processors: 1
      Total Number of Cores: 6
      L2 Cache (per Core): 256 KB
      L3 Cache: 12 MB
      Memory: 32 GB

</details>