@kavon 2024-08-30
In prior runs with minor tweaks to the algorithm, the `FlattenListFlatMap` benchmark has either not changed, or actually improved by 19%. So, I think it's overly sensitive and at some sort of arbitrary performance cliff; the result doesn't appear significant.

The only consistent improvements I've seen over multiple benchmark runs are:

1. the 5.2% code-size improvement in `ReversedCollections.o` with `-Osize`
2. the 20% performance improvement of ReversedBidirectional  with `-Osize`

They're related as (2) is part of the object file in (1), so my guess is the improvement is from not inlining as much into cold blocks.