@kavon 2024-09-03
@swift-ci Please smoke test compiler performance