@kavon 2024-09-04
@swift-ci Please test compiler performance