@kavon 2024-09-10
Results of the full compiler test:

```
# Summary for main full

**Regressions found (see below)**

- [Debug-batch](#debug-batch-80c582942450953)
    - [debug-batch brief](#debug-batch-brief-80c582942450953)
    - [debug-batch detailed](#debug-batch-detailed-80c582942450953)
- [Release](#release-80c582942450953)
    - [release brief](#release-brief-80c582942450953)
    - [release detailed](#release-detailed-80c582942450953)



<a name="debug-batch-80c582942450953">

# Debug-batch

<a name="debug-batch-brief-80c582942450953">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
time.swift-frontend.wall | 7.0s | 6.7s | -328.5ms | -4.69% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 215,839,632,990,503 | 215,691,186,209,614 | -148,446,780,889 | -0.07%
LLVM.NumLLVMBytesOutput | 4,882,677,480 | 4,881,768,268 | -909,212 | -0.02%
</details>

<a name="debug-batch-detailed-80c582942450953">

## debug-batch detailed
<details>
<summary>Regressed (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ModuleShadowCacheHit | 143,518 | 145,140 | 1,622 | 1.13% :no_entry:
AST.NumSourceLinesPerSecond | 10,909,204 | 11,433,270 | 524,066 | 4.8% :no_entry:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumProcessFailures | 47 | 46 | 0 | -1.06% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (362)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 10,650,348 | 10,646,945 | -3,403 | -0.03%
AST.ImportSetCacheMiss | 1,502,505 | 1,502,035 | -470 | -0.03%
AST.ImportSetFoldHit | 756,026 | 755,655 | -371 | -0.05%
AST.ImportSetFoldMiss | 746,479 | 746,380 | -99 | -0.01%
AST.ModuleShadowCacheMiss | 13,537 | 13,581 | 44 | 0.33%
AST.ModuleVisibilityCacheHit | 205,831,793 | 205,804,586 | -27,207 | -0.01%
AST.ModuleVisibilityCacheMiss | 853,817 | 853,743 | -74 | -0.01%
AST.NumASTBytesAllocated | 131,802,981,488 | 131,614,181,221 | -188,800,266 | -0.14%
AST.NumASTScopeExpansions | 18,723,776 | 18,713,803 | -9,973 | -0.05%
AST.NumASTScopeLookups | 24,085,105 | 24,073,911 | -11,194 | -0.05%
AST.NumDecls | 678,421 | 678,249 | -171 | -0.03%
AST.NumDependencies | 519,572 | 519,535 | -36 | -0.01%
AST.NumIncrementalDependencies | 108,821 | 108,821 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 639,053 | 639,004 | -49 | -0.01%
AST.NumLocalTypeDecls | 8,356 | 8,355 | 0 | -0.01%
AST.NumMacroPluginDependencies | 105 | 105 | 0 | 0.0%
AST.NumModuleLookupClassMember | 8,689 | 8,667 | -22 | -0.25%
AST.NumModuleLookupValue | 222,133,521 | 222,045,453 | -88,068 | -0.04%
AST.NumObjCMethods | 100,558 | 100,450 | -108 | -0.11%
AST.NumOperators | 1,005 | 998 | -7 | -0.7%
AST.NumPrecedenceGroups | 176 | 176 | 0 | 0.0%
AST.NumReferencedDynamicNames | 155 | 154 | -1 | -0.64%
AST.NumReferencedMemberNames | 35,258,341 | 35,243,820 | -14,521 | -0.04%
AST.NumReferencedTopLevelNames | 2,595,046 | 2,593,338 | -1,708 | -0.07%
AST.NumSourceBuffers | 801,907 | 801,512 | -395 | -0.05%
AST.NumSourceLines | 26,462,569 | 26,454,801 | -7,768 | -0.03%
AST.NumTotalClangImportedEntities | 2,425,868 | 2,422,068 | -3,800 | -0.16%
Frontend.MaxMallocUsage | 1,641,353,690,056 | 1,642,648,325,648 | 1,294,635,592 | 0.08%
Frontend.NumInstructionsExecuted | 215,839,632,990,503 | 215,691,186,209,614 | -148,446,780,889 | -0.07%
IRGen.IRGenRequest | 58,376 | 58,350 | -26 | -0.04%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 641,883 | 641,838 | -44 | -0.01%
IRModule.NumIRAliases | 438,937 | 438,801 | -136 | -0.03%
IRModule.NumIRBasicBlocks | 17,185,338 | 17,181,743 | -3,595 | -0.02%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 8,761,348 | 8,758,062 | -3,286 | -0.04%
IRModule.NumIRGlobals | 7,776,530 | 7,774,355 | -2,175 | -0.03%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 187,996,884 | 187,875,040 | -121,844 | -0.06%
IRModule.NumIRNamedMetaData | 235,026 | 234,922 | -104 | -0.04%
IRModule.NumIRValueSymbols | 16,657,312 | 16,651,745 | -5,567 | -0.03%
LLVM.NumLLVMBytesOutput | 4,882,677,480 | 4,881,768,268 | -909,212 | -0.02%
Parse.EvaluateIfConditionRequest | 483,956 | 483,944 | -12 | -0.0%
Parse.ExportedSourceFileRequest | 5,511 | 5,511 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 981,541 | 981,077 | -464 | -0.05%
Parse.NumIterableDeclContextParsed | 1,863,378 | 1,862,564 | -814 | -0.04%
Parse.ParseAbstractFunctionBodyRequest | 1,012,605 | 1,012,522 | -83 | -0.01%
Parse.ParseMembersRequest | 1,431,799 | 1,431,130 | -669 | -0.05%
Parse.ParseSourceFileRequest | 742,570 | 742,183 | -387 | -0.05%
Parse.ParseTopLevelDeclsRequest | 737,447 | 737,060 | -387 | -0.05%
SILGen.ASTLoweringRequest | 60,768 | 60,742 | -26 | -0.04%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 7,066,458 | 7,065,094 | -1,364 | -0.02%
SILModule.NumSILGenGlobalVariables | 258,351 | 258,403 | 52 | 0.02%
SILModule.NumSILGenVtables | 54,209 | 54,191 | -17 | -0.03%
SILModule.NumSILGenWitnessTables | 450,036 | 450,001 | -35 | -0.01%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 8,110,389 | 8,108,984 | -1,405 | -0.02%
SILModule.NumSILOptGlobalVariables | 261,407 | 261,461 | 54 | 0.02%
SILModule.NumSILOptVtables | 54,851 | 54,833 | -17 | -0.03%
SILModule.NumSILOptWitnessTables | 539,386 | 539,354 | -31 | -0.01%
SILOptimizer.ExecuteSILPipelineRequest | 299,086 | 298,956 | -130 | -0.04%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 685,589 | 685,413 | -176 | -0.03%
Sema.AbstractGenericSignatureRequest | 384,396 | 383,779 | -617 | -0.16%
Sema.AccessLevelRequest | 42,447,359 | 42,308,522 | -138,836 | -0.33%
Sema.ActorIsolationRequest | 92,484,446 | 92,163,240 | -321,206 | -0.35%
Sema.AllInheritedProtocolsRequest | 1,292,979 | 1,291,209 | -1,770 | -0.14%
Sema.AllMembersRequest | 361,964 | 361,920 | -44 | -0.01%
Sema.AnyObjectLookupRequest | 284 | 283 | -1 | -0.35%
Sema.ApplyAccessNoteRequest | 9,033,691 | 9,030,674 | -3,016 | -0.03%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 71,623 | 71,604 | -19 | -0.03%
Sema.AssociatedConformanceRequest | 1,037,772 | 1,037,505 | -267 | -0.03%
Sema.AttachedPropertyWrapperTypeRequest | 64,630 | 64,612 | -17 | -0.03%
Sema.AttachedPropertyWrappersRequest | 62,500,674 | 62,332,656 | -168,018 | -0.27%
Sema.AttachedResultBuilderRequest | 3,989,373 | 3,988,318 | -1,055 | -0.03%
Sema.BodyInitKindRequest | 165,600 | 165,561 | -39 | -0.02%
Sema.BraceHasReturnRequest | 2,053 | 2,053 | 0 | 0.0%
Sema.BreakTargetRequest | 15,890 | 15,890 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 820,333 | 820,104 | -229 | -0.03%
Sema.CanSynthesizeDistributedActorCodableConformanceRequest | 79 | 79 | 0 | 0.0%
Sema.CaptureInfoRequest | 1,713,401 | 1,712,891 | -509 | -0.03%
Sema.CheckDistributedFunctionRequest | 82 | 82 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 116,528 | 116,496 | -32 | -0.03%
Sema.CheckInconsistentAccessLevelOnImportSameFileRequest | 116,954 | 116,922 | -32 | -0.03%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 20,939 | 20,937 | -2 | -0.01%
Sema.CheckInconsistentSPIOnlyImportsRequest | 116,954 | 116,922 | -32 | -0.03%
Sema.CheckInconsistentWeakLinkedImportsRequest | 20,939 | 20,937 | -2 | -0.01%
Sema.CheckRedeclarationRequest | 4,794,606 | 4,793,705 | -901 | -0.02%
Sema.ClangCategoryLookupRequest | 217,742 | 217,025 | -717 | -0.33%
Sema.ClangDirectLookupRequest | 30,254 | 30,255 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 30,254 | 30,255 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 231,516 | 231,191 | -324 | -0.14%
Sema.ClosureEffectsRequest | 320,048 | 320,011 | -37 | -0.01%
Sema.ClosureHasResultExprRequest | 103,074 | 103,055 | -19 | -0.02%
Sema.CollectOverriddenDeclsRequest | 34,153,684 | 34,002,757 | -150,927 | -0.44%
Sema.CompareDeclSpecializationRequest | 1,951,066 | 1,949,098 | -1,968 | -0.1%
Sema.CompilerPluginLoadRequest | 105 | 105 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 2,207,184 | 2,205,121 | -2,063 | -0.09%
Sema.ConformanceAccessScopeRequest | 422,101 | 421,955 | -145 | -0.03%
Sema.ConformanceHasEffectRequest | 71 | 71 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 1 | 1 | 0 | 0.0%
Sema.ContinueTargetRequest | 4,965 | 4,961 | -3 | -0.07%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 38,976 | 38,980 | 3 | 0.01%
Sema.CustomAttrTypeRequest | 9,240 | 9,240 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 113,486 | 113,444 | -41 | -0.04%
Sema.DefaultAndMaxAccessLevelRequest | 284,765 | 284,679 | -86 | -0.03%
Sema.DefaultArgumentExprRequest | 251,610 | 251,520 | -90 | -0.04%
Sema.DefaultArgumentInitContextRequest | 1,355 | 1,353 | -1 | -0.11%
Sema.DefaultArgumentTypeRequest | 138,318 | 138,238 | -80 | -0.06%
Sema.DefaultDefinitionTypeRequest | 7,360 | 7,360 | 0 | 0.0%
Sema.DefaultInitializerIsolation | 2,648,306 | 2,647,808 | -498 | -0.02%
Sema.DefaultTypeRequest | 1,155,321 | 1,155,135 | -186 | -0.02%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 218,756,081 | 218,640,110 | -115,971 | -0.05%
Sema.DirectOperatorLookupRequest | 4,627,833 | 4,626,986 | -847 | -0.02%
Sema.DirectPrecedenceGroupLookupRequest | 1,758,184 | 1,757,428 | -756 | -0.04%
Sema.DistributedModuleIsAvailableRequest | 149 | 149 | 0 | 0.0%
Sema.DocCommentProvidingDeclRequest | 2,552 | 2,552 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 85,645,013 | 85,362,803 | -282,210 | -0.33%
Sema.EnumElementExprPatternRequest | 80 | 80 | 0 | 0.0%
Sema.EnumRawTypeRequest | 102,035 | 101,997 | -38 | -0.04%
Sema.EnumRawValuesRequest | 51,681 | 51,680 | -1 | -0.0%
Sema.ExistentialConformsToSelfRequest | 92,737 | 92,678 | -59 | -0.06%
Sema.ExpandAccessorMacros | 2,198,305 | 2,197,728 | -576 | -0.03%
Sema.ExpandBodyMacroRequest | 978,881 | 978,447 | -434 | -0.04%
Sema.ExpandChildTypeRefinementContextsRequest | 1,893,335 | 1,892,945 | -390 | -0.02%
Sema.ExpandExtensionMacros | 2,830,618 | 2,824,140 | -6,478 | -0.23%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 2,948 | 2,948 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 152,740,929 | 152,344,033 | -396,896 | -0.26%
Sema.ExpandPeerMacroRequest | 111,189,249 | 110,808,290 | -380,959 | -0.34%
Sema.ExpandPreambleMacroRequest | 978,881 | 978,447 | -434 | -0.04%
Sema.ExpandSynthesizedMemberMacroRequest | 660,078 | 659,851 | -227 | -0.03%
Sema.ExplicitCaughtTypeRequest | 62,349 | 62,338 | -11 | -0.02%
Sema.ExprPatternMatchRequest | 178,940 | 178,925 | -15 | -0.01%
Sema.ExtendedNominalRequest | 1,733,489 | 1,732,876 | -613 | -0.04%
Sema.ExtendedTypeRequest | 381,781 | 381,624 | -156 | -0.04%
Sema.ExternalMacroDefinitionRequest | 176 | 176 | 0 | 0.0%
Sema.FallthroughSourceAndDestRequest | 755 | 755 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 6,025,203 | 6,023,843 | -1,360 | -0.02%
Sema.FunctionOperatorRequest | 33,868 | 33,827 | -41 | -0.12%
Sema.GenericParamListRequest | 15,114,873 | 15,095,194 | -19,678 | -0.13%
Sema.GenericSignatureRequest | 6,773,388 | 6,766,996 | -6,392 | -0.09%
Sema.GetDestructorRequest | 60,812 | 60,756 | -56 | -0.09%
Sema.GetDistributedActorAsActorConformanceRequest | 5 | 5 | 0 | 0.0%
Sema.GetDistributedActorConcreteArgumentDecodingMethodRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 63 | 63 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 24 | 24 | 0 | 0.0%
Sema.GetDistributedMethodWitnessedProtocolRequirements | 0 | 0 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 14 | 14 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 21 | 21 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 2 | 2 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 2 | 2 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 12 | 12 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 82 | 82 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 141 | 141 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 101,149,469 | 100,810,778 | -338,691 | -0.33%
Sema.GlobalActorInstanceRequest | 5,228 | 5,232 | 4 | 0.08%
Sema.HasCircularInheritedProtocolsRequest | 793,607 | 792,096 | -1,511 | -0.19%
Sema.HasCircularRawValueRequest | 56,508 | 56,503 | -5 | -0.01%
Sema.HasDefaultInitRequest | 209,136 | 209,082 | -54 | -0.03%
Sema.HasDynamicCallableAttributeRequest | 75,683 | 75,667 | -16 | -0.02%
Sema.HasDynamicMemberLookupAttributeRequest | 1,203,803 | 1,203,049 | -754 | -0.06%
Sema.HasImportsMatchingFlagRequest | 1,447,363 | 1,446,589 | -774 | -0.05%
Sema.HasInitAccessorRequest | 12,013,480 | 11,975,485 | -37,995 | -0.32%
Sema.HasIsolatedSelfRequest | 111,452,192 | 111,077,374 | -374,818 | -0.34%
Sema.HasMemberwiseInitRequest | 137,135 | 137,114 | -20 | -0.01%
Sema.HasMissingDesignatedInitializersRequest | 41,143 | 41,132 | -11 | -0.03%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 23,968 | 23,954 | -14 | -0.06%
Sema.HasStorageRequest | 35,693,254 | 35,569,317 | -123,937 | -0.35%
Sema.HasUserDefinedDesignatedInitRequest | 209,142 | 209,087 | -54 | -0.03%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 8 | 8 | 0 | 0.0%
Sema.ImplicitKnownProtocolConformanceRequest | 1,610,373 | 1,606,113 | -4,260 | -0.26%
Sema.ImportDeclRequest | 139,446 | 139,398 | -47 | -0.03%
Sema.InferredGenericSignatureRequest | 506,477 | 505,737 | -740 | -0.15%
Sema.InheritedDeclsReferencedRequest | 7,452,420 | 7,433,038 | -19,382 | -0.26%
Sema.InheritedProtocolsRequest | 210,397 | 210,135 | -262 | -0.12%
Sema.InheritedTypeRequest | 944,709 | 943,860 | -849 | -0.09%
Sema.InheritsSuperclassInitializersRequest | 45,316 | 45,279 | -37 | -0.08%
Sema.InitAccessorPropertiesRequest | 140,081 | 140,058 | -23 | -0.02%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 392,629 | 392,490 | -139 | -0.04%
Sema.InterfaceTypeRequest | 51,031,567 | 50,899,014 | -132,553 | -0.26%
Sema.IsABICompatibleOverrideRequest | 711,858 | 711,747 | -110 | -0.02%
Sema.IsAccessorTransparentRequest | 1,379,413 | 1,379,194 | -219 | -0.02%
Sema.IsActorRequest | 5,117,699 | 5,107,749 | -9,950 | -0.19%
Sema.IsCCompatibleFuncDeclRequest | 0 | 0 | 0 | 0.0%
Sema.IsCallAsFunctionNominalRequest | 11,566 | 11,563 | -2 | -0.02%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 123,060 | 122,997 | -63 | -0.05%
Sema.IsDefaultActorRequest | 74,749 | 74,699 | -50 | -0.07%
Sema.IsDistributedActorRequest | 5,507,618 | 5,497,683 | -9,934 | -0.18%
Sema.IsDynamicRequest | 82,857,208 | 82,539,205 | -318,003 | -0.38%
Sema.IsFinalRequest | 82,479,852 | 82,160,823 | -319,028 | -0.39%
Sema.IsFunctionBodySkippedRequest | 3,536,749 | 3,535,474 | -1,275 | -0.04%
Sema.IsGetterMutatingRequest | 1,731,903 | 1,731,383 | -520 | -0.03%
Sema.IsImplicitlyUnwrappedOptionalRequest | 8,451,141 | 8,448,780 | -2,361 | -0.03%
Sema.IsNonUserModuleRequest | 283,194 | 283,220 | 26 | 0.01%
Sema.IsObjCRequest | 6,407,763 | 6,405,297 | -2,466 | -0.04%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 1,587,965 | 1,587,675 | -290 | -0.02%
Sema.IsSingleValueStmtRequest | 416,608 | 416,282 | -325 | -0.08%
Sema.IsStaticRequest | 3,239,462 | 3,238,105 | -1,357 | -0.04%
Sema.IsUnsafeRequest | 9,641,107 | 9,630,659 | -10,448 | -0.11%
Sema.LazyStoragePropertyRequest | 2,251 | 2,239 | -12 | -0.53%
Sema.LifetimeDependenceInfoRequest | 34,233,930 | 34,131,935 | -101,995 | -0.3%
Sema.LocalDiscriminatorsRequest | 1,258,635 | 1,258,556 | -79 | -0.01%
Sema.LocalTypeDeclsRequest | 116,954 | 116,922 | -32 | -0.03%
Sema.LookupAllConformancesInContextRequest | 7,598,184 | 7,569,530 | -28,654 | -0.38%
Sema.LookupConformanceInModuleRequest | 483,612,860 | 483,368,784 | -244,075 | -0.05%
Sema.LookupInModuleRequest | 23,560,583 | 23,544,163 | -16,420 | -0.07%
Sema.LookupInfixOperatorRequest | 205,150 | 205,118 | -32 | -0.02%
Sema.LookupIntrinsicRequest | 870 | 870 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 67,942 | 67,912 | -30 | -0.04%
Sema.LookupPrefixOperatorRequest | 506 | 505 | -1 | -0.2%
Sema.MacroDefinitionRequest | 114 | 114 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 4,171 | 4,171 | 0 | 0.0%
Sema.MemberwiseInitPropertiesRequest | 140,083 | 140,060 | -23 | -0.02%
Sema.ModuleImplicitImportsRequest | 20,939 | 20,937 | -2 | -0.01%
Sema.ModuleLibraryLevelRequest | 2 | 2 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 8,430,956 | 8,421,933 | -9,022 | -0.11%
Sema.NamedLazyMemberLoadSuccessCount | 140,822,997 | 140,649,669 | -173,328 | -0.12%
Sema.NamingPatternRequest | 510,024 | 509,625 | -398 | -0.08%
Sema.NeedsNewVTableEntryRequest | 2,056,820 | 2,056,452 | -368 | -0.02%
Sema.NumAccessorBodiesSynthesized | 668,536 | 668,493 | -43 | -0.01%
Sema.NumAccessorsSynthesized | 1,107,954 | 1,107,790 | -163 | -0.01%
Sema.NumConformancePathsRecorded | 2,605,618 | 2,605,233 | -385 | -0.01%
Sema.NumConformancesDeserialized | 22,334,665 | 22,274,400 | -60,265 | -0.27%
Sema.NumConstraintScopes | 150,047,632 | 150,017,544 | -30,088 | -0.02%
Sema.NumConstraintsConsideredForEdgeContraction | 10,976,504 | 10,977,121 | 616 | 0.01%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 215,635,578 | 215,168,988 | -466,590 | -0.22%
Sema.NumDeclsTypechecked | 6,891,232 | 6,889,914 | -1,317 | -0.02%
Sema.NumLazyIterableDeclContexts | 33,151,151 | 33,109,608 | -41,543 | -0.13%
Sema.NumLazyRequirementSignatures | 2,232,375 | 2,231,208 | -1,166 | -0.05%
Sema.NumLazyRequirementSignaturesLoaded | 1,107,292 | 1,105,811 | -1,481 | -0.13%
Sema.NumLeafScopes | 117,124,834 | 117,101,632 | -23,201 | -0.02%
Sema.NumRequirementMachineCompletionSteps | 57,687,255 | 57,561,740 | -125,515 | -0.22%
Sema.NumRequirementMachineUnifiedConcreteTerms | 5,014,366 | 5,008,079 | -6,287 | -0.13%
Sema.NumRequirementMachines | 4,420,295 | 4,409,934 | -10,361 | -0.23%
Sema.NumSwift6Errors | 31,739 | 31,737 | -2 | -0.01%
Sema.NumTypesDeserialized | 56,971,939 | 56,883,956 | -87,983 | -0.15%
Sema.NumUnloadedLazyIterableDeclContexts | 24,822,776 | 24,811,464 | -11,312 | -0.05%
Sema.ObjCCategoryNameMapRequest | 0 | 0 | 0 | 0.0%
Sema.ObjCInterfaceAndImplementationRequest | 4,796,534 | 4,795,045 | -1,489 | -0.03%
…